requirements.txt
ib_async==2.0.1
pandas>=2.0
psycopg2-binary==2.9.11
